[Platform][CLOUDGA-6229] add retries to subtasks
Summary:
YBM clusters often experience transient errors during creation and edit operations. These are mostly related to the SSH daemon crashing on GCP VMs, requiring a reboot. This changeset introduces retryability on a subtask level, allowing for post-failure actions before performing actual retries. This is supposed to help in certain scenarios such as
 - VM becoming unreachable during creation/edit (including ansible playbooks)
 - permanent errors during VM boot (see for example https://yugabyte.atlassian.net/browse/CLOUDGA-1759) where the instance has to be recreated

Changes in a nutshell:
1. Introduce a YBOpsRecoverableError in devops code and a counterpart RecoverableException in java. Devops code propagates recoverable errors using a custom exit code.
2. Add an onRetryableFailure method to ITask to implement any task specific actions before retrying a subtask (reboot/recreate the VM).
3. Add a 'reboot mode' argument to RebootServer (so that the instance can be rebooted using cloud provider SDK instead of using SSH).
4. Use the Retryable annotation to rerun failed subtasks in TaskExecutor, running their onRetryableFailure in between attempts.

Test Plan: Simulate a YBOpsRecoverableError at the end of creation/provisioning scripts, make sure the retryable failure hook runs as expected before a subtask retry attempt.

Reviewers: nsingh, daniel

Reviewed By: nsingh

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D19222

+++ managed/devops/opscli/ybops/cloud/common/method.py
@@ -21,7 +21,7 @@ import time
 import datetime
 
 from pprint import pprint
-from ybops.common.exceptions import YBOpsRuntimeError
+from ybops.common.exceptions import YBOpsRuntimeError, YBOpsRecoverableError
 from ybops.utils import get_path_from_yb, \
     generate_random_password, validate_cron_status, \
     YB_SUDO_PASS, DEFAULT_MASTER_HTTP_PORT, DEFAULT_MASTER_RPC_PORT, DEFAULT_TSERVER_HTTP_PORT, \
@@ -29,7 +29,7 @@ from ybops.utils import get_path_from_yb, \
 from ansible_vault import Vault
 from ybops.utils.ssh import wait_for_ssh, format_rsa_key, validated_key_file, \
     generate_rsa_keypair, scp_to_tmp, get_public_key_content, \
-    get_ssh_host_port, DEFAULT_SSH_USER
+    get_ssh_host_port, DEFAULT_SSH_USER, DEFAULT_SSH_PORT
 from ybops.utils import remote_exec_command
 
 
@@ -117,7 +117,7 @@ class AbstractInstancesMethod(AbstractMethod):
     """
     YB_SERVER_TYPE = "cluster-server"
     SSH_USER = "centos"
-    INSTANCE_LOOKUP_RETRY_LIMIT = 120
+    INSTANCE_LOOKUP_RETRY_LIMIT = 2
 
     def __init__(self, base_command, name, required_host=True):
         super(AbstractInstancesMethod, self).__init__(base_command, name)
@@ -272,7 +272,7 @@ class AbstractInstancesMethod(AbstractMethod):
             time.sleep(1)
             host_lookup_count += 1
 
-        raise YBOpsRuntimeError("Timed out waiting for instance: '{0}'".format(
+        raise YBOpsRecoverableError("Timed out waiting for instance: '{0}'".format(
             args.search_pattern))
 
     # Find the open ssh port and update the dictionary.
@@ -1202,13 +1202,16 @@ class ConfigureInstancesMethod(AbstractInstancesMethod):
                         args.itest_s3_package_path,
                         args.search_pattern, time.time() - start_time))
                 else:
-                    scp_to_tmp(
-                        args.package,
-                        self.extra_vars["private_ip"],
-                        self.extra_vars["ssh_user"],
-                        self.extra_vars["ssh_port"],
-                        args.private_key_file,
-                        ssh2_enabled=args.ssh2_enabled)
+                    if scp_to_tmp(
+                          args.package,
+                          self.extra_vars["private_ip"],
+                          self.extra_vars["ssh_user"],
+                          self.extra_vars["ssh_port"],
+                          args.private_key_file,
+                          ssh2_enabled=args.ssh2_enabled):
+                        raise YBOpsRecoverableError(
+                            f"[app] Failed to copy package {args.package} to {args.search_pattern}")
+
                     logging.info("[app] Copying package {} to {} took {:.3f} sec".format(
                         args.package, args.search_pattern, time.time() - start_time))
 
@@ -1216,13 +1219,15 @@ class ConfigureInstancesMethod(AbstractInstancesMethod):
                 ybc_package_path = args.ybc_package
                 if os.path.isfile(ybc_package_path):
                     start_time = time.time()
-                    scp_to_tmp(
-                        ybc_package_path,
-                        self.extra_vars["private_ip"],
-                        self.extra_vars["ssh_user"],
-                        self.extra_vars["ssh_port"],
-                        args.private_key_file,
-                        ssh2_enabled=args.ssh2_enabled)
+                    if scp_to_tmp(
+                          ybc_package_path,
+                          self.extra_vars["private_ip"],
+                          self.extra_vars["ssh_user"],
+                          self.extra_vars["ssh_port"],
+                          args.private_key_file,
+                          ssh2_enabled=args.ssh2_enabled):
+                        raise YBOpsRecoverableError(f"[app] Failed to copy package "
+                                                    f"{ybc_package_path} to {args.search_pattern}")
                     logging.info("[app] Copying package {} to {} took {:.3f} sec".format(
                         ybc_package_path, args.search_pattern, time.time() - start_time))
 
@@ -1612,6 +1617,11 @@ class RebootInstancesMethod(AbstractInstancesMethod):
     def __init__(self, base_command):
         super(RebootInstancesMethod, self).__init__(base_command, "reboot")
 
+    def add_extra_args(self):
+        super().add_extra_args()
+        self.parser.add_argument("--use_ssh", action='store_true', default=False,
+                                 help="Use 'sudo reboot' instead of cloud provider SDK")
+
     def callback(self, args):
         host_info = self.cloud.get_host_info(args)
         if not host_info:
@@ -1624,15 +1634,24 @@ class RebootInstancesMethod(AbstractInstancesMethod):
         if ssh_user is None:
             ssh_user = DEFAULT_SSH_USER
 
-        self.extra_vars.update(get_ssh_host_port(host_info, args.custom_ssh_port))
-        self.extra_vars.update({"ssh_user": ssh_user})
-        rc, stdout, stderr = remote_exec_command(
-                                self.extra_vars["ssh_host"],
-                                self.extra_vars["ssh_port"],
-                                self.extra_vars["ssh_user"],
-                                args.private_key_file,
-                                'sudo reboot', ssh2_enabled=args.ssh2_enabled)
-        self.wait_for_host(args, False)
+        if args.use_ssh:
+            use_default_port = not self.update_open_ssh_port(args)
+            self.extra_vars.update(get_ssh_host_port(host_info, args.custom_ssh_port,
+                                                     default_port=use_default_port))
+            self.extra_vars.update({"ssh_user": ssh_user})
+            rc, stdout, stderr = remote_exec_command(
+                                    self.extra_vars["ssh_host"],
+                                    self.extra_vars["ssh_port"],
+                                    self.extra_vars["ssh_user"],
+                                    args.private_key_file,
+                                    'sudo reboot', ssh2_enabled=args.ssh2_enabled)
+            if rc:
+                raise YBOpsRecoverableError(f"Failed to connect to {args.search_pattern}")
+
+            self.wait_for_host(args, False)
+        else:
+            extra_vars = get_ssh_host_port(host_info, args.custom_ssh_port)
+            self.cloud.reboot_instance(args, [DEFAULT_SSH_PORT, extra_vars["ssh_port"]])
 
 
 class RunHooks(AbstractInstancesMethod):
